Cleanups, skeleton for event support; test script.
