+++ src/gentl/README.md
@@ -1,0 +1,13 @@
+GenTL interface for Aravis.
+
+Function signatures are defined by EMVA; the header `GenTL_v1_6.h` is included as distributed by EMVA (BSD-license + prohibition of modification).
+
+Implementation of those functions is split into modules indicated by function prefixes:
+
+* `GC`: GeniCam (system)
+* `TL`: transport layer
+* `IF`: interface
+* `Dev`: device
+* `DS`: data stream
+* `Event`
+
